+++ package.json
@@ -24,12 +24,13 @@
   ],
   "scripts": {
     "dev": "vitepress --open",
-    "build": "vitepress build",
     "serve": "vitepress serve",
-    "fetch-api": "node fetchApi.js",
-    "generate-sitemap": "node generateSitemap.js",
-    "pnpm": "pnpm install",
-    "start": "vitepress serve"
+    "build-only": "vitepress build",
+    "generate-sitemap": "node src/generateSitemap.js",
+    "build": "run-s build-only generate-sitemap"
+  },
+  "dependencies": {
+    "fs-extra": "^11.2.0"
   },
   "devDependencies": {
     "@tailwindcss/typography": "^0.5.10",
@@ -39,6 +40,7 @@
     "axios": "^1.4.0",
     "feed": "^4.2.2",
     "gray-matter": "^4.0.3",
+    "npm-run-all": "^4.1.5",
     "sitemap": "^7.0.0",
     "tailwindcss": "^3.4.1",
     "vitepress": "^1.0.0-rc.41",
@@ -60,8 +62,5 @@
   "engines": {
     "nvm": "22",
     "node": "^22.13.0"
-  },
-  "dependencies": {
-    "fs-extra": "^11.2.0"
   }
 }
